+++ Model/PaymentTypeManagement/Cpt.php
@@ -13,18 +13,21 @@ class Cpt extends AbstractPaymentTypeManagement
     ) {
         $orderAmount = $this->helperData->mapMagentoAmountToPaylineAmount($payment->getOrder()->getGrandTotal());
         $responseAmount = $response->getAmount();
-
         if ($responseAmount == $orderAmount) {
             return true;
-        } else {
-            $message = __(
-                'ERROR for order ; payment gateway amount %1 does not match order amount %2.',
-                $response->getAmount(),
-                $this->helperData->mapMagentoAmountToPaylineAmount($payment->getOrder()->getGrandTotal())
-            );
-            $payment->setAmountToCancel($response->getAmount());
-            $this->handlePaymentCanceled($payment->setData('payline_in_error', true), $message);
-            return false;
         }
+
+
+
+        $message = __(
+            'Payment gateway amount %1 does not match order amount %2.',
+            $this->helperData->mapPaylineAmountToMagentoAmount($response->getAmount()),
+            $payment->getOrder()->getGrandTotal()
+        );
+
+        $payment->setAmountToCancel($response->getAmount());
+        $this->handlePaymentCanceled($payment->setData('payline_in_error', true), $message);
+        return false;
+
     }
 }